+++ packages/zosjobs/__tests__/__unit__/CancelJobs.unit.test.ts
@@ -43,6 +43,9 @@ describe("Cancel Jobs unit tests", () => {
         "job-correlator": "mycorrelator",
         "exec-started": '2024-01-02T15:57:58.350Z',
         "exec-ended": '2024-01-02T15:58:00.600Z',
+        "exec-member": 'SYS1',
+        "exec-system": 'SYS1',
+        "exec-submitted": '2024-01-02T15:58:00.600Z'
     };
     describe("Positive tests", () => {
         it("should allow users to call cancelJob with correct parameters", async () => {