+++ packages/zosjobs/src/doc/response/IJob.ts
@@ -49,6 +49,27 @@ export interface IJob {
     "exec-ended": string;
 
     /**
+     * exec-member
+     * @type {string}
+     * @memberof IJob
+     */
+    "exec-member": string;
+
+    /**
+     * exec-submitted
+     * @type {Date}
+     * @memberof IJob
+     */
+    "exec-submitted": string;
+
+    /**
+     * exec-system
+     * @type {String}
+     * @memberof IJob
+     */
+    "exec-system": string;
+
+    /**
      * The primary or secondary JES subsystem.
      * If this value is null, the job was processed by the primary subsystem.
      * @type {string}